abstract pool buttons
